Refactor move partial_model decorator
From schema to utils.config_templates
