+++ lib/galaxy/schema/__init__.py
@@ -1,21 +1,14 @@
-from collections.abc import Iterable
-from copy import deepcopy
 from datetime import datetime
 from enum import Enum
 from typing import (
-    Any,
-    Callable,
     Optional,
-    TypeVar,
     Union,
 )
 
 from pydantic import (
     BaseModel,
-    create_model,
     Field,
 )
-from pydantic.fields import FieldInfo
 
 
 class BootstrapAdminUser(BaseModel):
@@ -116,43 +109,3 @@ class PdfDocumentType(str, Enum):
 class APIKeyModel(BaseModel):
     key: str = Field(..., title="Key", description="API key to interact with the Galaxy API")
     create_time: datetime = Field(..., title="Create Time", description="The time and date this API key was created.")
-
-
-T = TypeVar("T", bound="BaseModel")
-
-
-# TODO: This is a workaround to make all fields optional.
-#       It should be removed when Python/pydantic supports this feature natively.
-# https://github.com/pydantic/pydantic/issues/1673
-def partial_model(
-    include: Optional[list[str]] = None, exclude: Optional[list[str]] = None
-) -> Callable[[type[T]], type[T]]:
-    """Decorator to make all model fields optional"""
-
-    if exclude is None:
-        exclude = []
-
-    def decorator(model: type[T]) -> type[T]:
-        def make_optional(field: FieldInfo, default: Any = None) -> tuple[Any, FieldInfo]:
-            new = deepcopy(field)
-            new.default = default
-            new.annotation = Optional[field.annotation or Any]  # type:ignore[assignment]
-            return new.annotation, new
-
-        if include is None:
-            fields: Iterable[tuple[str, FieldInfo]] = model.model_fields.items()
-        else:
-            fields = ((k, v) for k, v in model.model_fields.items() if k in include)
-
-        return create_model(
-            model.__name__,
-            __base__=model,
-            __module__=model.__module__,
-            **{
-                field_name: make_optional(field_info)
-                for field_name, field_info in fields
-                if exclude is None or field_name not in exclude
-            },
-        )  # type:ignore[call-overload]
-
-    return decorator
