+++ lib/galaxy/schema/schema.py
@@ -33,7 +33,6 @@ from typing_extensions import (
     TypedDict,
 )
 
-from galaxy.schema import partial_model
 from galaxy.schema.bco import XrefItem
 from galaxy.schema.fields import (
     DecodedDatabaseIdField,
@@ -49,6 +48,7 @@ from galaxy.schema.types import (
     RelativeUrl,
 )
 from galaxy.tool_util_models.tool_source import FieldDict
+from galaxy.util.config_templates import partial_model
 from galaxy.util.hash_util import HashFunctionNameEnum
 from galaxy.util.sanitize_html import sanitize_html
